chore: extract normalizeTokenAddress to utils
